Perform renames in TASVideo
