+++ src/Plugins.Win32.TASVideo/2xSAI.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
-#include "Types.h"
-#include "GBI.h"
+#include "Types.hpp"
+#include "GBI.hpp"
 
 static inline s16 GetResult1(u32 A, u32 B, u32 C, u32 D, u32 E)
 {
