+++ src/Plugins.Win32.TASVideo/CMakeLists.txt
@@ -5,38 +5,38 @@ SPDX-License-Identifier: GPL-2.0-or-later
 ]===]
 
 add_library(Mupen64RR.Plugins.Win32.TASVideo MODULE
-    "VI.h"
-    "Types.h"
-    "unified_combiner.h"
+    "VI.hpp"
+    "Types.hpp"
+    "unified_combiner.hpp"
     "stdafx.h"
-    "S2DEX2.h"
-    "S2DEX.h"
-    "RSP.h"
+    "S2DEX2.hpp"
+    "S2DEX.hpp"
+    "RSP.hpp"
     "resource.h"
-    "RDP.h"
-    "OpenGL.h"
-    "N64.h"
-    "L3DEX2.h"
-    "L3DEX.h"
-    "L3D.h"
-    "gSP.h"
-    "glN64.h"
-    "gDP.h"
-    "GBI.h"
-    "FrameBuffer.h"
-    "F3DWRUS.h"
-    "F3DPD.h"
-    "F3DEX2.h"
-    "F3DEX.h"
-    "F3DDKR.h"
-    "F3D.h"
-    "DepthBuffer.h"
-    "CRC.h"
-    "convert.h"
-    "Config.h"
-    "Combiner.h"
-    "3DMath.h"
-    "2xSAI.h"
+    "RDP.hpp"
+    "OpenGL.hpp"
+    "N64.hpp"
+    "L3DEX2.hpp"
+    "L3DEX.hpp"
+    "L3D.hpp"
+    "gSP.hpp"
+    "glN64.hpp"
+    "gDP.hpp"
+    "GBI.hpp"
+    "FrameBuffer.hpp"
+    "F3DWRUS.hpp"
+    "F3DPD.hpp"
+    "F3DEX2.hpp"
+    "F3DEX.hpp"
+    "F3DDKR.hpp"
+    "F3D.hpp"
+    "DepthBuffer.hpp"
+    "CRC.hpp"
+    "convert.hpp"
+    "Config.hpp"
+    "Combiner.hpp"
+    "3DMath.hpp"
+    "2xSAI.hpp"
 
     "VI.cpp"
     "Textures.cpp"
