+++ src/Plugins.Win32.TASVideo/CRC.hpp
