+++ src/Plugins.Win32.TASVideo/Config.hpp
