+++ src/Plugins.Win32.TASVideo/DepthBuffer.hpp
