+++ src/Plugins.Win32.TASVideo/F3D.cpp
@@ -1,12 +1,12 @@
 #include "stdafx.h"
-#include "glN64.h"
-#include "F3D.h"
-#include "N64.h"
-#include "RSP.h"
-#include "RDP.h"
-#include "gSP.h"
-#include "gDP.h"
-#include "GBI.h"
+#include "glN64.hpp"
+#include "F3D.hpp"
+#include "N64.hpp"
+#include "RSP.hpp"
+#include "RDP.hpp"
+#include "gSP.hpp"
+#include "gDP.hpp"
+#include "GBI.hpp"
 
 void F3D_SPNoOp(u32 w0, u32 w1)
 {
