+++ src/Plugins.Win32.TASVideo/F3DDKR.cpp
@@ -1,13 +1,13 @@
 #include "stdafx.h"
-#include "glN64.h"
-#include "F3D.h"
-#include "F3DDKR.h"
-#include "N64.h"
-#include "RSP.h"
-#include "RDP.h"
-#include "gSP.h"
-#include "gDP.h"
-#include "GBI.h"
+#include "glN64.hpp"
+#include "F3D.hpp"
+#include "F3DDKR.hpp"
+#include "N64.hpp"
+#include "RSP.hpp"
+#include "RDP.hpp"
+#include "gSP.hpp"
+#include "gDP.hpp"
+#include "GBI.hpp"
 
 void F3DDKR_DMA_Mtx(u32 w0, u32 w1)
 {
