+++ src/Plugins.Win32.TASVideo/FrameBuffer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.h"
+#include "Types.hpp"
 #include "Textures.h"
 
 struct FrameBuffer
