+++ src/Plugins.Win32.TASVideo/GBI.cpp
@@ -1,22 +1,22 @@
 #include "stdafx.h"
-#include "glN64.h"
-#include "GBI.h"
-#include "RDP.h"
-#include "RSP.h"
-#include "F3D.h"
-#include "F3DEX.h"
-#include "F3DEX2.h"
-#include "L3D.h"
-#include "L3DEX.h"
-#include "L3DEX2.h"
-#include "S2DEX.h"
-#include "S2DEX2.h"
-#include "F3DDKR.h"
-#include "F3DWRUS.h"
-#include "F3DPD.h"
-#include "Types.h"
+#include "glN64.hpp"
+#include "GBI.hpp"
+#include "RDP.hpp"
+#include "RSP.hpp"
+#include "F3D.hpp"
+#include "F3DEX.hpp"
+#include "F3DEX2.hpp"
+#include "L3D.hpp"
+#include "L3DEX.hpp"
+#include "L3DEX2.hpp"
+#include "S2DEX.hpp"
+#include "S2DEX2.hpp"
+#include "F3DDKR.hpp"
+#include "F3DWRUS.hpp"
+#include "F3DPD.hpp"
+#include "Types.hpp"
 #include "resource.h"
-#include "CRC.h"
+#include "CRC.hpp"
 
 u32 uc_crc, uc_dcrc;
 char uc_str[256];
