+++ src/Plugins.Win32.TASVideo/GBI.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.h"
+#include "Types.hpp"
 
 // Microcode Types
 #define F3D 0
