+++ src/Plugins.Win32.TASVideo/L3D.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.h"
+#include "Types.hpp"
 
 #define L3D_LINE3D 0xB5
