+++ src/Plugins.Win32.TASVideo/L3DEX.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.h"
+#include "Types.hpp"
 
 void L3DEX_Line3D(u32 w0, u32 w1);
 void L3DEX_Init();
