+++ src/Plugins.Win32.TASVideo/L3DEX2.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.h"
+#include "Types.hpp"
 
 #define L3DEX2_LINE3D 0x08
