+++ src/Plugins.Win32.TASVideo/OpenGL.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "gSP.h"
+#include "gSP.hpp"
 
 struct GLVertex
 {
