+++ src/Plugins.Win32.TASVideo/RSP.cpp
@@ -1,16 +1,16 @@
 #include "stdafx.h"
-#include "glN64.h"
-#include "OpenGL.h"
-#include "RSP.h"
-#include "RDP.h"
-#include "N64.h"
-#include "F3D.h"
-#include "3DMath.h"
-#include "VI.h"
-#include "Combiner.h"
-#include "FrameBuffer.h"
-#include "DepthBuffer.h"
-#include "GBI.h"
+#include "glN64.hpp"
+#include "OpenGL.hpp"
+#include "RSP.hpp"
+#include "RDP.hpp"
+#include "N64.hpp"
+#include "F3D.hpp"
+#include "3DMath.hpp"
+#include "VI.hpp"
+#include "Combiner.hpp"
+#include "FrameBuffer.hpp"
+#include "DepthBuffer.hpp"
+#include "GBI.hpp"
 #include <hqx.h>
 
 RSPInfo RSP;
