+++ src/Plugins.Win32.TASVideo/S2DEX2.cpp
@@ -1,15 +1,15 @@
 #include "stdafx.h"
-#include "OpenGL.h"
-#include "S2DEX.h"
-#include "S2DEX2.h"
-#include "F3D.h"
-#include "F3DEX.h"
-#include "F3DEX2.h"
-#include "GBI.h"
-#include "gSP.h"
-#include "gDP.h"
-#include "RSP.h"
-#include "Types.h"
+#include "OpenGL.hpp"
+#include "S2DEX.hpp"
+#include "S2DEX2.hpp"
+#include "F3D.hpp"
+#include "F3DEX.hpp"
+#include "F3DEX2.hpp"
+#include "GBI.hpp"
+#include "gSP.hpp"
+#include "gDP.hpp"
+#include "RSP.hpp"
+#include "Types.hpp"
 
 void S2DEX2_Init()
 {
