+++ src/Plugins.Win32.TASVideo/S2DEX2.hpp
