+++ src/Plugins.Win32.TASVideo/Textures.cpp
@@ -1,15 +1,15 @@
 #include "stdafx.h"
-#include "OpenGL.h"
+#include "OpenGL.hpp"
 #include "Textures.h"
-#include "GBI.h"
-#include "RSP.h"
-#include "gDP.h"
-#include "gSP.h"
-#include "N64.h"
-#include "CRC.h"
-#include "convert.h"
-#include "2xSAI.h"
-#include "FrameBuffer.h"
+#include "GBI.hpp"
+#include "RSP.hpp"
+#include "gDP.hpp"
+#include "gSP.hpp"
+#include "N64.hpp"
+#include "CRC.hpp"
+#include "convert.hpp"
+#include "2xSAI.hpp"
+#include "FrameBuffer.hpp"
 
 TextureCache cache;
