+++ src/Plugins.Win32.TASVideo/Types.hpp
