+++ src/Plugins.Win32.TASVideo/VI.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.h"
+#include "Types.hpp"
 
 struct VIInfo
 {
