+++ src/Plugins.Win32.TASVideo/gDP.cpp
@@ -1,17 +1,17 @@
 #include "stdafx.h"
-#include "glN64.h"
-#include "N64.h"
-#include "GBI.h"
-#include "RSP.h"
-#include "gDP.h"
-#include "gSP.h"
-#include "Types.h"
-#include "convert.h"
-#include "OpenGL.h"
-#include "CRC.h"
-#include "FrameBuffer.h"
-#include "DepthBuffer.h"
-#include "VI.h"
+#include "glN64.hpp"
+#include "N64.hpp"
+#include "GBI.hpp"
+#include "RSP.hpp"
+#include "gDP.hpp"
+#include "gSP.hpp"
+#include "Types.hpp"
+#include "convert.hpp"
+#include "OpenGL.hpp"
+#include "CRC.hpp"
+#include "FrameBuffer.hpp"
+#include "DepthBuffer.hpp"
+#include "VI.hpp"
 
 gDPInfo gDP;
