+++ src/Plugins.Win32.TASVideo/gSP.cpp
@@ -1,18 +1,18 @@
 #include "stdafx.h"
-#include "glN64.h"
-#include "Types.h"
-#include "RSP.h"
-#include "GBI.h"
-#include "gSP.h"
-#include "gDP.h"
-#include "3DMath.h"
-#include "OpenGL.h"
-#include "CRC.h"
+#include "glN64.hpp"
+#include "Types.hpp"
+#include "RSP.hpp"
+#include "GBI.hpp"
+#include "gSP.hpp"
+#include "gDP.hpp"
+#include "3DMath.hpp"
+#include "OpenGL.hpp"
+#include "CRC.hpp"
 #include <string.h>
-#include "convert.h"
-#include "S2DEX.h"
-#include "VI.h"
-#include "DepthBuffer.h"
+#include "convert.hpp"
+#include "S2DEX.hpp"
+#include "VI.hpp"
+#include "DepthBuffer.hpp"
 #include "resource.h"
 
 #define gSPFlushTriangles()                                                                                            \
